Updated with way to generate planner only testcases

+++ test-case-generator.py
@@ -38,7 +38,7 @@ PERTURB_TEMPLATE ='{"path" : <PATH>, "perturb-loc" : "<PERTURB_LOC>", "perturb" 
 CONFIGS = ['amcl-kinect', 'amcl-lidar', 'mrpt-lidar', 'mrpt-kinect', 'aruco-camera']
 
 # The set of configuration-based perturbations
-PERTURBS = ["node", "sensor", "nodeandsensor"]
+PERTURBS = ["node", "sensor", "nodeandsensor", "none"]
 
 # The set of light perturbations: lights mean include lights, lightsonly means ignore other perturbations, nolights means don't perturb the light
 LIGHT_PERTURBS = ["lights", "lightsonly", "nolights", "nolights", "nolights"]
@@ -150,37 +150,40 @@ def get_config_info(only_lights=False, existing=None):
     configurations.add(configuration)
     return configuration, config, perturbs, light_perturbs, src, tgt, utility, path
 
-def generate_icse_test_cases_planner(src, tgt, path='.'):
+def generate_icse_test_cases_planner(src, tgt, path='.', dry_run=False, allcombos=False):
     """
     Generates a case for all instances unders src -> tgt
     i.e., all utilities and configurations
     For the case where ony the planner is run
     Randomly choosing perturb and perturb location
     """
+    newly_generated = 0
     perturbs = random.choice(PERTURBS)
     light_perturbs = random.choice(LIGHT_PERTURBS)
     if light_perturbs == 'lightsonly':
         light_perturbs='lights'
-
-    for c in CONFIGS:
-        for u in UTILITIES:
-            route = inst_db.get_path(src, tgt, u + "-" + c)
-            if len(route) == 0:
-                print("Warning, cannot find path between %s and %s under %s for preference %s" %(src,tgt,c,u))
-                continue
-            ploci = random.randint(1, len(route)-2)
-            ploc = route[ploci]
-            for a in range(1,MAX_RECONFIGS):
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-1' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, False, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-1' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False,False, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, fix_path=True, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False, fix_path=True, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-1-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, False, fix_path=True, ploc=ploc)
-                gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-1-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False,False, fix_path=True, ploc=ploc)
-               
-
+    for perturbs in PERTURBS:
+        for light_perturbs in ["lights", "nolights"]:
+            for c in CONFIGS:
+                if perturbs != "none" or (light_perturbs=="lights" and c == "aruco-camera"):
+                    for u in UTILITIES:
+                        route = inst_db.get_path(src, tgt, u + "-" + c)
+                        if len(route) == 0:
+                            print("Warning, cannot find path between %s and %s under %s for preference %s" %(src,tgt,c,u))
+                            continue
+                        ploci = 0; #random.randint(1, len(route)-2)
+                        ploc = route[ploci]
+                        for a in range(1,MAX_RECONFIGS):
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-1' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, False, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-1' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False,False, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, fix_path=True, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False, fix_path=True, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-cost-1-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,True, False, fix_path=True, ploc=ploc, dry=dry_run)
+                            newly_generated = newly_generated + gen_planner_case('c', '%s-%s-%s-%s-%s-%s-%s-nocost-1-fp' %(c,perturbs,light_perturbs,src,tgt,u,a), path, c, perturbs, light_perturbs, src, tgt, u, route,a,False,False, fix_path=True, ploc=ploc, dry=dry_run)
+                       
+    return newly_generated
 
 
 def generate_icse_test_cases(src, tgt, compose_file, path='.', genab=True):
@@ -250,8 +253,11 @@ def generate_test_case(compose_file, path = ".", only_lights=False, cases=['c'],
     for case in cases:
         gen_case(case,id,compose_file,path,config,perturbs,light_perturbs,src,tgt,utility,route,corridor,perturb_sleep)
 
-def set_up_directories(path,id,case,start,ros = True):
+def set_up_directories(path,id,case,start,ros = True,dry=False):
     test_directory = "%s/testcase-%s-%s/" %(path,id,case)
+    if os.path.exists(test_directory): return test_directory, True
+    if dry:
+        return test_directory, False
     os.makedirs(test_directory + start)
     os.makedirs(test_directory + "logs/prism")
     os.makedirs(test_directory + "logs/alloy")
@@ -264,10 +270,14 @@ def set_up_directories(path,id,case,start,ros = True):
         os.makedirs(test_directory + "roslogs")
         os.chmod(test_directory + "roslogs", stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
         touch(test_directory + "roslogs/.INCLUDE_ME")
-    return test_directory
+    return test_directory, False
 
-def gen_planner_case(case, id, path, config, perturbs, light_perturbs, src, tgt, utility, route, num_adaptations=1,consider_cost=False,balanced_utility=True,fix_path=False, ploc=None):
-    test_directory = set_up_directories(path, id, case, "params", False)
+def gen_planner_case(case, id, path, config, perturbs, light_perturbs, src, tgt, utility, route, num_adaptations=1,consider_cost=False,balanced_utility=True,fix_path=False, ploc=None, dry=False):
+    test_directory, existed = set_up_directories(path, id, case, "params", False, dry)
+    if existed: 
+        return 0
+    elif dry: 
+        return 1
     readyjson = get_ready(config,src,tgt,utility,case)
     with open("%s/params/start.json" %test_directory, 'w') as f:
         f.write(readyjson)
@@ -280,11 +290,11 @@ def gen_planner_case(case, id, path, config, perturbs, light_perturbs, src, tgt,
         f.write(perturbjson)
 
     shutil.copyfile('icse.properties', '%s/params/properties.properties' %test_directory)
-
+    return 1
 
 def gen_case(case,id, compose_file, path,config,perturbs,light_perturbs,src,tgt,utility,route,corridor,perturb_sleep,num_adaptations=1,consider_cost=False, balanced_utility=True, fix_path=False):
     # create directory structure
-    test_directory = set_up_directories(path,id,case, "start")
+    test_directory, existed = set_up_directories(path,id,case, "start")
 
     # create ready msg
     readyjson = get_ready(config,src,tgt,utility, case)
@@ -358,6 +368,8 @@ parse_args.add_argument('-d', "--existing", type=str, help='Do not duplicated th
 parse_args.add_argument('-t', '--tasks', type=str, help="The file containing the lists of tasks (src,target) that should be generated")
 parse_args.add_argument('--icse', action="store_true", help="Generate test cases for ICSE paper")
 parse_args.add_argument('--planner-only', action="store_true", help='Generate scenarios to run the plan only')
+parse_args.add_argument("--append", action="store_true", help='Append missing test cases to whatever exists, rather than overwriting')
+parse_args.add_argument("--dry-run", action="store_true", help='Just report how many cases would be generated, but do not generate them')
 parse_args.add_argument('num', type=int, help="number of tests (per machine) to generate")
 
 args = parse_args.parse_args()
@@ -393,13 +405,19 @@ if args.tasks is not None:
 
 if args.icse:
     if args.planner_only:
+        created = 0
         for m in range(args.machines):
-            os.makedirs('icse-planner%s' %m)
+            if args.append and not os.path.exists('icse-planner%s' %m):
+                os.makedirs('icse-planner%s' %m)
+            elif not args.append:
+                os.makedirs('icse-planner%s' %m)
             for i in range(args.num):
-                generate_icse_test_cases_planner(tasks[0]['src'], tasks[0]['tgt'],path='icse-planner%s' %m)
+                created = created + generate_icse_test_cases_planner(tasks[0]['src'], tasks[0]['tgt'],path='icse-planner%s' %m, dry_run = args.dry_run)
                 del tasks[0]
-            shutil.copyfile('run-icse-tests.sh', 'icse-planner%s/run-all-tests.sh' %m)
-            os.chmod('icse-planner%s/run-all-tests.sh' %m, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
+            if not args.dry_run:
+                shutil.copyfile('run-icse-tests.sh', 'icse-planner%s/run-all-tests.sh' %m)
+                os.chmod('icse-planner%s/run-all-tests.sh' %m, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
+            print("Created %s new test cases" %created)
 
     else:
         for m in range(args.machines):
